Cleaned up the prover and prover rewriter.
Removed the use of data_expression() as indicator of a failed
result by adding an explicit boolean. This avoids constructing a default
term very often.

Made it possible that the rewrite prover also applies to terms not of
type bool. This was now explicitly excluded. All tests succeed, showing
that this application does not cause harm.

Changed some progress messages from debug1 to debug, such that
they are printed with the -d flag.

Removed a number of dedicated functions to get the k'th argument of
an application, and to determine the number of arguments of an
application. There are standard functions for this.

Passed quite a number of arguments by reference, instead of by value.

Except that the rewriter now also rewrites non boolean expressions
using the prover rewriter, the behaviour should be unchanged.