Largely resolved #1595: multi_actions are not normalized.
Changed the lts libraries such that they only contain normalized
multiactions. When reading in .aut, .lts and .fsm format multiactions
are ordered. The same happens when they are generated.

This means that sometimes a multiactoin b|a is read, but it is
internally represented as a|b.

The only tool that has not been adapted is lps2ltsdeprecated.
For this the files explore.cpp and explore.h still need to be
adapted.