Added a sorting routine on aterm_lists.
Added a sorting routine on aterm lists that sorts on the
stack, unless the list is very long.

This is done to replace an ugly routine in timed_multi_actions
that used a static set for performance reasons.

The generic routine should probably be adapted such that it
can use an arbitary sorting function. Now it sorts on addresses.