Repaired #1584 by removing unlucky typecheck clause.
The routine parse_state_formula_specification contained an
unlucky line:

data::typecheck_data_specification(result.data());

which would check the data in a modal formula independently of
the data in a linear process specification for which it was
intended.

It turns out that by removing this line a .mcf file with data
can be combined without problems with the .lps file, and
type checking of the combined lists of actions and data types
turn out to work fine.

This resolves #1584.