Further changes to structured ouput: use ; to delimit actions
Using a space to delimit actions makes parsing somewhat more difficult, 
since pretty printed actions with multiple parameters contain a space 
themselves.
The return code of ltscompare is now 0 again, regardless of the result 
(consistent with the rest of the toolset). The result of the comparison 
is printed to stdout. The structured output resembles YAML.