Added extended tests to illustrate a type checking problem.
Complex right hand sides of action rename rules in an action rename
specification do not seem to be type checked properly, leading to residual errors.
Rules that were not properly typechecked were already in the test cases,
but there was no explicit check on whether a renamed linear process was
well typed.