Added an include path in AddMCRL2Binary.cmake for sylvan.
Added a statement

   set(ARG_INCLUDE ${ARG_INCLUDE} ${CMAKE_SOURCE_DIR}/3rd-party/sylvan/src)

to AddMCRL2Binary.cmake such that the header tests can find the sylvan
include. This resolves a failing header test.

As I am not fluent in cmake code, I do not know whether this is the
"natural" solution. Maybe a better statement can be found to achieve
the same.