[mcrl2ide] Changed the buttons of the addEditPropertyDialog
The buttons are now as follows:
 - Save and Parse: Saves the property to the properties folder and then parses it.
 - Save and Close: Saves the property to the properties folder and then closes the dialog.
 - Close: Closes the dialog

This change makes it no longer necessary to save properties in the temporary folder.
Updating the propertiesDock while using the addEditPropertyDialog is now handled using signals sent from the fileSystem to enable these updates while the dialog is open.

This is a partial solution for issue #1612