[mcrl2ide] Whenever you close the add/edit property dialog while there are unsaved changes, it will now notify you about this
You then have the option to save the changes (and then close the dialog), don't save the changes (and then close the dialog) or close the dialog.
This behaviour is consistent with the main window when there are unsaved changes in the main specification.

This is a partial solution for issue #1612

Also changed the notification text slightly (for the specification) to emphasize that the changes are unsaved.