Strengthened the representant generator to generate lambda terms.
The representant generator generates given a sort s a term of that sort.
Up till now no for function sorts, a representant was only generated
if a function would exist matching that sort. This has now been extended
by generating a lambda term if such a function does not exist.
For instance for the sort   Nat#Real -> Nat now the representant

     lambda x0:Nat, x1:Real.0

is generated.

Added a simple test to check the representant generator for function sorts.