Resolve a problem hiding hidden actions.
It turned out that

des (0,2,3)
(0,"a",1)
(2,"a",1)

would not be transformed properly when applying

 ltsconvert test.aut --tau=b test2.aut

This was due to the fact that labels were put in a map mapping them
to tau, but this map was only implicitly applied when searching
for reachable states or doing a bisimulation reduction.

Now actions are renamed explicitly in ltsconvert, before applying
any reduction.

Furthermore, the map mapping action labels to tau labels has been
replaced by a set of actions that must be considered hidden, as
this is the more natural notion.