Added as an example the small 2x2x2 rubick's cube
The mCRL2 specification of this small 2x2x2 rubick's cube allows
it to generate the state space completely, being less than 90M states.
This can for instance be used to find the minimal sequence of turns
to solve the puzzle. The enclosed run.py file does this concretely
for one scrambled cube. It is easy to add the colors of a scrambled
cube and let the system solve the cube, although this is not
particularly efficient, as it simply generates all possible rotations
until a situation is reached that corresponds to a solved cube.

Tempting remaining work could be to derive
a cookbook using which based on the colors of the cube the
optimal rotation could be done to solve it.