Optimised the lineariser a little bit.
In mCRL2 it is possible to specify

   ... a. X()

where X is defined by another process Y

   X() = Y()

and it is even possible that Y is again defined by Z, etc.

If the Y and Z would also have been used elsewhere after an
action, the lineariser could generate separate processes
for X, Y and Z tripling the number of summands, and als
unnecessarily increasing the number of states in the process.
This has been adapted in the sense that only the last process
name, i.e. Y or Z above in such a sequence will be used to
generate processes from.

This may have some benefit for processes using such constructions,
especially for the mCRL2 processes generated by the Cordis-UML
translator.

All tests succeed.