Implemented summand groups in lpsreach
*) It turns out that there is a complication when joining summands in a
summand group. The PhD thesis of Jeroen Meijer contains an algorithm
that suggests this has been solved by using a special value in the LDDs
of transition relations to indicate that a parameter needs to be left
unchanged. It is unknown whether this is implemented in the Sylvan
function relprod, and what that special value would be.