Changed the interface of the application contructor.
The application constructor

     application(h, t1,...,tn, f)

had an implementation that required a number of copies of its
arguments. This turned out to be expensive. The implementation has
been adapted such that this copying is avoided.

At the same time the interface has been changed slightly.

Up till now the function f was only applied to the arguments t1,...,tn,
and not to the header h. From now on f will also be applied to the
header h. In case this is not desired, and there were two use cases
in the toolset for that, an optional argument false can be passed along.
I.e.

      application(h, t1,....,tn, f, false)

will prevent the function f from being applied to h.

In the description above, t1,...,tn are given by two iterators
first and last.

All tests, except for a few header tests belonging to pbesses and
symbolic reachability succeed.