Made abstraction(const aterm&) explicit.
The constructor transforming an aterm into an abstraction
was not explicit.

There were several places where implicit transformations to
abstractions took place, including some that were erroneous.

Added a comment in a file in utilities, which is made dependent
on aterms, which is not allowed.

There are still a number of failing header checks, mainly related
to recent work on lpsreach. This requires repairs.