Replace an inefficient function in print.h.
Replaced an inefficient function in print.h, that would
remove some type casting functions in data_expressions
numerous times. Using a lambda function this could be
made a little bit more efficient.

Besides that added a note to the use of aterm_list in
utilities. The utilities library may not depend on the
aterm library. This particular function in the current
form must be removed.

All tests succeed, except headertests related to container_utility,
due to the problem above, and headertests related to the new
symbolic algorithms.