Enabled the possibility to implement functions in the data types explicitly in C++
Added to each data specification file the option to indicate for each
function whether it would be implemented in C++ or as a rewrite rule.
Furthermore, made it explicit whether a function would internal, i.e.,
only used within the tools, not being available to a user, or whether
it would be external.

The motivation of this change is that an efficient implementation of
function updates requires a C++ defined function. This work is
preparatory for that.

No function is currently implemented in C++.

Some, but not all functions used in the data types are now generated
from the data.spec files, instead of being explicitly enumerated in
the type checker.

Part of the implementation is borrowed from the 64 bit implementation
of built in numbers.

All tests succeed as it stands (except four header tests).