Let the typechecker use the declared datatypes.
Up till now the typechecker had an explicit list of
functions that it was using. Now it derives this list
from the types in the data.spec files, for bool,
numbers and lists. The other types still need to
be made dependent on the ....spec files

All tests, except two header tests succeed.