Allow one projection function for multiple arguments in code generation.
This fixes clearly incorrect code in the code generation for standard data types in the case where the same projection function was used to project arguments in different positions.
This code was clearly broken and even syntactically incorrect, so probably never even used. I am leaving it in here since we are processing user input that may change.

The fix properly generates one case for each position that needs to be projected.

The issue was spotted by Jan Friso Groote