Rewrote function updates by using some c++ implemented rewrite rules.
Replaced the exponential rules for function updates by rules with
linear complexity. For this it was necessary to use a few c++
implemented rules. So, both the jitty and jittyc rewriter have
been adapted to allow the inclusion of such c++ defined rules.
They can simply be declared in a xxxx.spec file defining a datatype
by writing behind a declaration of a function or a mapping that
it is implemented_by_code. Furthermore, a file in
libraries/data/include/mcrl2/data/detail/xxx.h must be made containing
the implementation of this constructor or mapping.