Some minor improvements.
Added two rewrite rules for non stable function updates.
Under normal circumnstances these might not be used, but
for weird specs they may make life substantially easier.

Improved the code of is_finite_sort, which made an unnecessary
copy of a vector with constructors.

Adapted the jittyc rewriter such that it does not throw away
the rewrite cpp file when using the --debug flag (when compiled in
releas mode). This file is required for seeing in the profile which
compiled rewrite function is being executed.