Resolved bug #1644
With help of Wieger Wesselink bug #1644 namely that bound fixed point variables
in modal formulas were not properly renamed, is solved. The problem was
that the renaming operation on variables in a builder was not written as
an apply but as an operator (). But that meant that the operator was not
applied, not renaming variables.

A test has been added to catch this problem in case it occurs again.
All tests succeed.