Resolved bug 1645 by adding internal data functions to the typechecker.
The typechecker made a distinction between external functions that
can be used in an mCRL2 specification and internal functions. This
change has been introduced recently.

However, sometimes .lps and .pbes specs are typechecked, and these
contain internal functions. These always start with a "@" symbol
hence it is syntactically impossible to use such functions in
textual specifications.

As a solution all internally defined functions are now added to the
typechecker such that it will recognize them. This resolves bug #1645.

All tests succeed.