Repaired a nasty renaming error.
When performing a renaming in the lineariser it was
possible that two parameter lists that were joined
could have the same parameter name. This occurred in particular
when joining lists where the name of one list would contain
names that would occur in the other list with an extra number as
postfix. This was a serious error, which is already in the toolset
for a long time, and it is remarkable that this was not detected
earlier.