Resolved bug #1648 multiactions wrongly renamed.
Hiding of multi-actions took place by hiding the individual
actions, but replacing them in the wrong order. So, if c and d were
not hidden, then c|d would be "hidden" to d|c.
Then it would be checked whether the result was not the same as the
original from which it was concluded that the action was completely
hidden. Hence any multiaction was erroneously renamed to tau.

The stupid check that multiactions were changed is now replaced by
a check that the result equals tau, resolving this problem.

All tests succeed.