Prepared lps::multi_action to be merged with process::timed_multi_action
The classes lps::multi_action and process::timed_multi_action
essentially do exactly the same, namely storing an action list
together with a  expression of type real, which can be undefined.

Having two similar notions is not wise. For instance they are
transformed into each other whenever a .lts is read or written to disk.

In this commit the lps::multi_action is made essentially equal to
process::timed_multi_action by deriving from it.

The next step is to completely replace lps::multi_action by
process::timed_multi_action.

It is conceivable to rename process::timed_multi_action to
process::multi_action, simply because it is shorter.

All tests succees, although due to the different nature of
lps::multiaction, which was modifiable (but not anymore), and timed_multi_action,
which is non modifiable subtle issues may arise.

Timings show that this change appears to have little influence, and
optimisations by removing explicit conversions between the two types
still have to be applied, which will be most noticable in the trace and
.lts format.