[mcrl2-gui] Fixed some teamcity build errors
Apparently, QProcess::startDetached() (the non-static version) was introduced in Qt 5.10, while some builds on teamcity use Qt 5.9. Now the non-static startDetached is used if Qt >= 5.10 and the static version (as it was before the previous commit) is used in case Qt < 5.10 and not Windows. In case Qt < 5.10 and Windows we simply do not add the menu item.
Relates to issue #1652