Provided make_aterm_... variants
Provided make_term_list, make_aterm_appl and make_aterm_int
functions that can create a term at a particular address.
The purpose of this is that in a setting with protection
through hashsets, this save a large number of entries and
removals of variables from the protection hash_set, as
the result is directly constructed at a particular, already
protected, address.

There are not yet test for this construct.

e Please enter the commit message for your changes. Lines starting