Removed a bug in the trace generation in the new state space generator.
When generating a trace, the method add_action(a) moves the current
position in the trace to the end of the trace. Adding a new action
and state must be done by:

   add_action(a)
   set_state(s)

to extend the trace with an action a going to state s.

In the state space generator it was done in the wrong sequence:

   set_state(s)
   add_action(a)

This led to replacing the last state of the current trace by s and
adding a dangling action a after it. The new trace library noticed
the problem, as traces were not well formed.