Adapted the names for equivalences in liblts_failures_refinement.h
The names for various equivalences in liblts_failures_refinement were
trace, failures and failures_divergence. These names occurred in the
mcrl2/lts namespace causing conflicts with the class trace.

This is confusing the Microsoft compilers.

Renamed the equivalences by putting ..._preorder behind them to avoid
a conflict.

It is questionable whether enumerated types for very particular purposes
should live in major namespaces. Maybe it is better to move this
enumerated type inside the appropriate class. That has not been done.

Tests are still running, but they should not cause issues.