Avoid a lot of copying in the creation of case functions.
A lot of time in the generation of case functions was spent in creating and destroying temporary aterm_appl and aterm_list variables, and copying lists around.
This was observed while profiling a case where we were creating clustered summands with roughly 50,000 elements.

With this change, we are still not able to cluster this particular example, but the time is now predominantly spent in the data library.