Slightly adapted the jitty rewriter.
Try to figure out where unnecessary reference counter increments are
taking place. This leads to some remove of code. In particular,
rewrite_aux_function_symbol should be applied to an application,
and not to a data_expression. I will apply this to the master also,
as it leads to a substantial code cleanup.