Attempt to use a vector as a stack in the jitty rewriter.
Used an atermpp vector as a stack the jitty rewriter. This leads
to quite some complications, and is actually slower
with reference counting than the version without such a vector.
With a std::vector is does not crash, but with an atermpp::vector
it tend to crash regularly, most likely due to the fact that
multiple processes seem to be involvedn in garbage collection.