[mcrl2-gui] path argument values passed to a process are no longer surrounded by quotes
This fixes issue #1654
Since commit 38111b658e3da33ac3ff77cd736f9b88910fa059 a QStringList is used instead of a QString to supply the arguments. QProcess then uses the argument list implementation of some OSs, which removes the need for the surrounding quotes. To still be compatible with OSs that can can only represent arguments as a single string (such as Windows), QProcess adds the quotes automatically for paths with spaces. Since we added these quotes manually, such argument values got too many quotes.
A parameter was added to still add these quotes when used to log the command in console, so that the command logged still corresponds to the command run.