Atermpp::deque now appears to work properly.
Managed to generate a state space with 1 process using an
atermpp::deque. Did not yet try a state space with multiple processes.
Biggest problem is now the utilities::indexed_set, which used a
std::vector and a std::deque, and hence they cannot be used to
store terms by one thread that are then read by another thread.