Added an atermpp::indexed_set
The atermpp::indexed set is the same as the
mcrl2::utitilies::indexed_set, except that it uses an atermpp::deque
instead of a std::deque. This means that different parallel threads
can write into this set when using the protection_set mechanism
without issues. If a std::deque is used, threads remain owner of
terms written to the deque. If a thread is destroyed, there is no
protection for the elements. In a atermpp::deque the creator of the
deque is the owner of the terms, and a thread releases ownership.