Added the marking protection of an unordered map.
Took care that we can use an atermpp::unordered_map with
automatic protection of the elements in the map.
This was tricky, because an unordered map is using pairs,
and some members of these pairs contain or are aterms, and
need marking, and other members of such pairs are non aterms,
and should not be marked during mark and sweep garbage collection.

Generation of 1394-fin and abp work, with 1 and four threads
in lps2lts_parallel, but there are still various moments where
the system is unreliable.