The jittyc rewriter can now be used also for parallel rewriting.
All the rewriters have been adapted such that a rewriter
is now copied as a whole, instead of being handed over via a
shared pointer. This caused the jittyc rewriter to work.
Compiling multiple rewriters simultaneously is not really
possible.

The jittyc rewriter is relatively slow because it is very
exensively using protection sets, as it is not adapted to
new aterm library.