All tests seem to pass again, with more extensive checking of type conversions.
Several coding errors were exposed, where terms were incorrectly
down_casted. This has been improved.

One remarkable observation is that an aterm_balanced_tree derives
from an aterm_appl, but actually, for trees of size one, it has the
type of the element put in the tree. This element can be an aterm_int
or term_list, meaning that the proper type of an aterm_balanced_tree
should be an aterm.

This should be repaired.