Repaired a failing random test.
pbes2bool failed due to the fact that atermpp::down_cast<propositional_variable_instantiation>
was applied to a term that was not a
propositional_variable_instantiation. Actually, it was: X0(3, 3, 3) &&
Zpos_0_b(1, 3, 3, 3, 3, 3) || Zneg_0_b(1, 3, 3, 3, 3, 3).

In the code it appeared that such cases were skipped more or less
accidentially. An explicit if now does the same.