Added an escape for strange behaviour of the gcc compiler.
The gcc compiler in optimizing mode appears to call a destructor
for which a destructor has not been called. It is unclear what the
source of this information is. But this leads to removing
terms from the protection set that have never been inserted.
As an escape, this situation is now detected and ignored.
Hopefully, this can be removed in due time.