Made the index for locally generated function symbols local.
Up till now the index for generated function symbol was global.
Due to an increment that was not part of a critical section, it was
possible that various function symbol generators used the same prefix.
This meant that if the global index was reset by one function symbol
generator, it would also be reset in others, causing havoc, as
newly generated variables would overlap with old ones.