A solution for the crash with parallel processes and creation hooks.
The crash in parallel processes was due to the fact that
a function symbol or a variable was created with a sequence number.
The number was looked up. Then a garbage collection could come by,
deleting any occurrence of that variable with that number, and
then, erroneously a variable would be created again with an old
non existing sequence number.

The solution is to assign the sequence number while creating a
function symbol or a variable.

This must also be applied at other occurrences where sequence numbers
are assigned, such as process id's and bes variables.

Extensive experiments must now show whether this is working. Initial
tests are promising.