Disabled the on_create functions.
The on create functions exist for variables, function symbols,
process identifiers, bes variables and pbes variables. They are
invoked and call an insertion in the index trait set. However,
the inserted number is not used. Instead the numbered items are
constructed while they are created.

It seems that in the current form the "on_create" functionality is
not required and can be removed. They do cause problems in the parallel
branch, as indices that are created around garbage collections can
yield wrong indices.

As it stands the on_create functionality should probably be removed.
This will speed up all creations of aterm objects.

The on_delete functionality is crucial and must remain. It removes
objects from the index_trait index sets when the objects are removed
during garbage collection.