Disabled on_create_.... functions.
The on_create_... functions do not appear to do
anything useful, but they can interfere with parallelly triggered
garbage collection. Therefore, they have been outcommented.
As the on_create functionality is not used anywhere, it appears
to me that it can be removed completely.

The on_destroy_... functions and functionality is essential,
as it used to remove the index of a particular object from the
index_traits set, such that it can be reused later. It is called
when that particular object is removed during garbage collection.