Extended the vector_too_large exception to all rewrite constructs.
In jitty a rewrite_stack_too_large exception is thrown if the
vector containing the rewrite stack must be resized. This means
that the current term that is being rewritten will be rewritten from
the start with a bigger stack. If this happens
the substitution sigma, which may have changed when rewriting
with binders, must also be brought back to the initial shape.
Code to do so has been added. Furthermore, an additional
rewrite function

     rewrite(result, term, sigma)

that places the result in its first argument has been added, as
this avoids having a temporary variable that requires protection
on the stack.

As it stands, various tests fail, but I have to move to another
computer...