Reimplemented how constructors with an extra index are implemented.
Some classes have constructors where a unique index is put as an extra
argument. These are variable, function_symbol, process_identifier,
boolean_variable and a propositional_variable_instantiation.

This code was not resilient against parallel construction, as the
index was sometimes as there already existed constructors with the same
index. Then a garbage collection could remove all constructors with this
index, including the removal of the index for this constructor.
Subsequently, the old, already retrieved index, was used to construct
a new constructor with an index not in the index administration. All
subsequent equal constructors would get a new index, meaning that
the same object woud exist with two indices.

The solution was to surround creating the constructor and the index
surrounded by a shared lock. Then no garbage collection could pass by
in the intermediate time.

All tests succeed, except that they randomly fail at the end
when apparently sometimes aterms are removed from the protection set,
whereas it appears that the protection set does not exist anymore.