Prepared the compiling rewriter to be translated and removed a bug in the naming of .cpp files.
First a problem in naming the .cpp files of the compiling rewriter has
been addressed. The name was made using the process ID and the address
of the compiling rewriter, more precisely the "this" pointer.
Unfortunately, in long test runs, such as lps2lts_test.cpp where
rewriters are created/compiled continuously, many files where created.
Sometimes these would be the same, namely if the rewriter in use
would be positioned in memory at the same address. Then the system
would dynamically load the old rewriter. This rewriter would contain
pointers to old addresses. It is not yet clear to me why the old
files are not overwritten by the newly generated ones, but apparently
this is a feature of the compiler/linker.

This problem has been resolved by adding a time to the name of the
generated .cpp files. As compiling takes at least a second, and often
longer, and it is not possible that two compiling rewriters in the same
thread at the same address are compiled with different content, this
should be safe. It is not completely watertight, though.

The problem above is not only due to the parallel variant of the
toolset, but also for the sequential one. Maybe this solution should
be transferred to that one as well, but for some reason this problem
is not occurring there.

For the remainder the compiling rewriter has been prepared to use
a atermpp::vector as a rewrite stack, although this stack is not
being used. All rewrite functions in the compiling rewriter are now
of type void rewr_i_j(data_expression& result, ....) instead of
the previous type data_expression rewr_i_j(...).

All 5000+ tests succeed, including all tool tests.