Adapted the jittyc rewriter to avoid inefficient protection.
Adapted the jittyc rewriter such that it uses the atermpp::vector
as stack, in such a way avoiding too many insertions and deletions
in the protection set.

The jittyc rewriter in the sequential master branch requires 10 seconds
to generate the state space of 1394-fin.lps. The as it stands the
lps2lts requires in the parallel branch 36 seconds.

The time for two threads in lps2lts_parallel is 29s, and for four
threads 19s. With 8 threads (on a 4 core +4 hyperthreads) intel mac
15s are required. With one thread this is: 56s.

As it stands these parallel times are disappointing, and it needs
to be figured out what the reason for this is.

All tests succeed on the mac.