Substantial speed increase of the compiling rewriter.
Inspired by work on the parallel branch of the toolset,
it became obvious that the very often used rewrite of a constant
term was inefficient as it required the creation of a term
"application()", which was not used at all. By putting all normal forms
of constant terms into a separate vector, this can be avoided.
This yields a more than 25% speed increase of 1394-fin.

All tests succeed.