Moved thread_local declaration to a .cpp file.
On the Mastodont, apparently declaring thread_local variables in
include files lead to having a duplication when dynamically loading
compiled files.

Concretely, the thread administration recoreds were registered twice
when using the compiling rewriter. The concrete effect was a duplication
of the m_busy_flags, and the term administration became confused as to
which m_busy_flags had to be used when.

By moving the thread_local declaration to a .cpp file the problem
disappears.