Made the busy and forbidden flag accessible via the rewriters.
As some form of hack, the busy and forbidden flag have been
put inside the rewriters such that they can be accessed more
easily. This helps somewhat in performance, but this trick
should be removed when the access of thread local variables
in C++ will have become faster than it is now, which no doubt
will happen.