Merged lps2lts_parallel with lps2lts
The code used in the experimental tool lps2lts_parallel has
largely been transferred to lps2lts. This means that lps2lts can
in principle run in parallel (except that it does not have the option
--threads, and that is the only difference now with lps2lts_parallel.

lps2lts now spawns off one thread to do the state space generation
(previously it used the main thread). In lps2lts using the --threads
flag multiple threads to generate the state space can be spawn.

Single thread state space generation works well, and all tests succeed.

Within lps2lts_parallel all flags from lps2lts can now be used, but
they do not always work. For instance --cached with multiple threads
is not yet thread safe. This is work for the next phase.

As it stands performance and memory usage of the new state
space generation are slightly worse than that of lps2lts of the
release. For jitty this is 5% and for jittyc it is 30%.
It can however be expected that this will improve. The same
holds for memory usage, which is not expected, as the new
aterm library uses 20% memory per aterm.

All tests, including the jittyc and all tooltests succeed.

There are still some debug statements in the code.