Adapted make_list_forward and make_list_backward
The functions make_list_forward and make_list_backward have
been adapted to avoid making intermediate aterms, which are now
expensive due to the use of protection sets for variables.

New variants have been added that store their results directly
in a target variable, instead of delivering the results on the stack
in an explicitly protected anonymous variable.

As it stands intermediate results are still stored in an
MCRL2_SPECIFIC_STACK_ALLOCATOR of which the positions are
still explicitly protected in protection sets. This needs
to be adapted too.

All tests succeed in debug mode.