Allowed function in term constructions with a finite number of arguments.
In constructing at term by make_term_appl or term_appl, it is now
possible to provide as arguments constant functions. These functions
are evaluated when the arguments of a term are needed. This prevents
to have to store a term in a separate protected place in memory.

For instance with f an unary term:

      term_appl(f, [arg1&](void){ return arg1; });

has the same effect as writing

      term_appl(f, arg1);

but   term_appl(f, [arg1&](void){ return term_appl(g, arg1); });
will only construct the term g(arg1) when it is needed as an
argument.

It is also possible to store the result in place in the function, as in

      term_appl(f, [arg1&](aterm& result){ result=arg1; });

All tests succeed in debug mode.