Some further performance improvements
When checking whether two terms in the hashtable are equal,
first the arguments and then the function symbols are checked.
This has been changed, as there are only a few function symbols,
and this means that checking equality of function symbols succeeds
relatively often.

Now a thread safe version of indexed sets is being used, instead
of guarding the use of indexed sets by one mutex. As it stands the
use of indexed sets is the performance bottleneck for processes with
many states that can easily be calculated, using multiple processes.
This needs to be improved further, most likely by replacing the
implementation of indexed sets.

Furthermore, each process now has a local buffer of newly found states,
instead of pushing each found state directly into the todo buffer. This
prevents locking the mutex for the todo buffer too often.

Put a static guard around the message "Explored the maximum number ...
of states", as otherwise it was printed quite often, even in a single
threading setting.

All sets in debug mode succeed.