The indexed sets are now parallel.
Made indexed sets parallel as they are a bottleneck for
parallel generation of statespaces where a large number of states can be generated
in a short timespan.

The idea is that every thread reserves a large number of indices and
hands these out. This has as disadvantage that the numbers that are
handed out are not always contiguous. For single threads the behaviour
is unchanged.

All tests succeed (at least on a mac).