Replaced the builder framework.
Replaced the builder framework. Builders use functions apply and
update to traverse through objects and deliver changed objects.
The updates are unchanged. "Apply" had the general type

     resulttype apply(const objecttype& t)

This has been replaced by apply functions of the shape:

     template <class RESULTTYPE>
     void (RESULTTYPE& result, const objecttype& t)

The old variant would implicitly create a result on the stack,
requiring always a protection by an insert and a remove in a protection
set. This is now relatively very costly.

The new variant allows to construct the result into the variable
result, and as such avoid one protection of a term variable.
Especially the new functions make_....(result, arguments) can
be used for this. For instance make_and(result, left, right)
constructs a term and(left, right) in result, without requiring
any explicit protection.

Instead of terms left and right, it is also possible to use
a function  (void){ ... return left; }, or a function i
(resulttype& r){ .... r=left; } that calculate the left and
right arguments, while the new and function is constructed.
This also does not require any explicitly protected terms, and
is therefore quite efficient. Internally, the construction
of lists and balanced trees use this feature.

This transformation was necessary to make a competitive implementation
of all programs using the builder framework, in particular
pbessolve/pbes2bool.

All tests succeed (again) on the mac. The compiler warns that
c++20 features are used. Maybe these should be removed.