Added a few namespaces required on linux/c++17
On the mac -- for reasons I do not understand --
a function without its namespace could be used. This was declared
a c++20 feature by the clang compiler. On linux it was clearly stated
that a namespace should be added. This caused compilation on TeamCity
to go astray on Ubuntu.

This should now be resolved.