Changed a condition under which resizing of tables was initiated.
There was a conter m_count_until_resize which triggered a resize
of hashtables if the value would become 0. In a parallel setting this
value could also be decremented twice before being compared. This
would lead to a value -1, after which the value would be decremented,
but never reach 0. From this point on hashtables were not resized
anymore, making the toolset sluggish when there would be many terms.

The behavior with the above error was very unpredictable. If the counter
would become negative early in a program run, the tables would be small
and running time draconic. If the problem would occur late, then
performance could be very reasonable. In different test runs, I saw
a difference of a factor 20 at times.