By writing to unprotected terms in lambda functions, performance is improved.
In make term  of the shape make_term(result, function_symbol, args, transformer)

the transformer is most efficient if it is of type

     [&](aterpp::unprotected_aterm& result, const termtype& argument){ ....  result=...}

If result is a protected aterm, the compiler calls the busy_forbidden
protocol when assigning to result. But as result is assigned within the
scope of constructing a term, the busy forbidden was already called, and
the nested call is redundant.