Addressed bug #1668.
Remnants of very old code assumed that there were only a limited
number of arguments in terms. This of course is not true under all
circumstances. The code of the nfs_array has been adapted.

There is evidence that this code could be removed/refreshed at more
places. The assert is removed and cannot cause havoc.
It is unlikely that this affected the release.