Slightly adapted the data_rewriter for pbesses.
The data rewriter for pbesses has slightly been adapted
to be more friendly regarding protection sets. Results
are not passed via the stack, put put in place directly.
A temporary vector to store arguments has been replaced
by lambda functions that put the arguments, in this case
parameters of a PBES variable, directly in place.
This has a noticable performance improvement of appr.
10 percent when deadlock checking with 1394-fin with
the compiling rewriter.